+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change log
 All notable changes to this project will be documented in this file
 
+## 1.1.0 - 2016-07-19
+Added
+- more pages
+- form that submits via a POST request
+
 ## 1.0.0 - 2016-07-18
 Added
 - basic server functionality
