+++ templates/form.html
@@ -1,5 +1,6 @@
+<h2>Leave me a message here</h2>
+
 <form method="POST" action="/thanks">
-  <h2>Leave me a message here</h2>
-  <textarea></textarea>
+  <input name="message">
   <button type="submit">Send</button>
 </form>
